Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioworklet-interface/audioworkletglobalscope-timing-info.https-expected.txt

CONSOLE MESSAGE: line 52: TypeError: undefined is not an object (evaluating 'context.audioWorklet.addModule')

FAIL 
      Test currentTime and currentFrame in AudioWorkletGlobalScope
     TypeError: undefined is not an object (evaluating 'context.audioWorklet.addModule')
